requirements.txt
torch==2.4.1
bittensor==9.0.0
matplotlib==3.9.2
python-dotenv==1.0.1
rich==13.9.2
safetensors==0.4.5
numpy==2.0.1
transformers==4.51.3
wandb==0.18.3
datasets==3.0.1
smart-open[s3]==7.0.5
accelerate
torchvision